requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
distlib==0.3.8
Django==5.0.6
django-allauth==0.63.2
filelock==3.14.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
googlemaps==4.10.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
Markdown==3.6
oauthlib==3.2.2
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
